np.int is deprecated. Replaced with np.int32
check at file WordMatching.py, line 91

+++ WordMatching.py
@@ -88,7 +88,7 @@ def get_resulting_string(mapped_indices: np.array, words_estimated: list, words_
     number_of_real_words = len(words_real)
     for word_idx in range(number_of_real_words):
         position_of_real_word_indices = np.where(
-            mapped_indices == word_idx)[0].astype(np.int)
+            mapped_indices == word_idx)[0].astype(np.int32)
 
         if len(position_of_real_word_indices) == 0:
             mapped_words.append(WORD_NOT_FOUND_TOKEN)
